rename crypto file to metamaskCrypto
