+++ examples/notes/src/metamaskCrypto.ts
